Various fixes/changes
