+++ src/bedrock_server_manager/core/server/process_mixin.py
@@ -527,235 +527,5 @@ class ServerProcessMixin(BedrockServerBaseMixin):
     def get_process_info(self) -> Optional[Dict[str, Any]]:
         """
         Gets resource usage information (PID, CPU%, Mem MB, Uptime) for the running server process.
-        CPU% is delta-calculated using instance-specific state and requires psutil.
-        The first call after (re)start or a long pause in monitoring for this server instance
-        will report 0.0% CPU for this instance.
-
-        Returns:
-            A dictionary with process info, or None if server not running or psutil unavailable.
-
-        Raises:
-            ResourceMonitorError: If psutil is required but unavailable (though typically returns None),
-                                  or for unexpected errors during psutil operations.
         """
-        if not PSUTIL_AVAILABLE:
-            self.logger.warning(
-                "'psutil' package not found. Cannot get detailed process info."
-            )
-            raise ResourceMonitorError("'psutil' is required for get_process_info.")
-
-        self.logger.debug(
-            f"Attempting to get process info for server '{self.server_name}' (using delta CPU)..."
-        )
-
-        bedrock_psutil_process: Optional["psutil_for_types.Process"] = None
-        pid: Optional[int] = None
-        # Normalized path to the server's directory (used for CWD check on Linux)
-        target_server_dir_norm = os.path.normpath(os.path.abspath(self.server_dir))
-
-        try:
-            # --- Find the target Bedrock Process ---
-            if self.os_type == "Linux":
-                screen_session_name = f"bedrock-{self.server_name}"
-                parent_screen_pid: Optional[int] = None
-                for proc_iter in psutil.process_iter(["pid", "name", "cmdline"]):
-                    try:
-                        if proc_iter.info[
-                            "name"
-                        ] == "screen" and screen_session_name in " ".join(
-                            proc_iter.info.get("cmdline", [])
-                        ):
-                            parent_screen_pid = proc_iter.info["pid"]
-                            break
-                    except (
-                        psutil.Error,
-                        TypeError,
-                    ):  # psutil.Error covers NoSuchProcess, AccessDenied, ZombieProcess
-                        continue
-                if not parent_screen_pid:
-                    self.logger.debug(
-                        f"Screen session '{screen_session_name}' not found for '{self.server_name}'."
-                    )
-                    return None
-
-                # Find the bedrock_server child of the screen session
-                found_child = False
-                for proc_iter in psutil.process_iter(
-                    ["pid", "ppid", "name", "cwd", "status"]
-                ):
-                    try:
-                        proc_info = proc_iter.info
-                        if (
-                            proc_info["name"] == self.bedrock_executable_name
-                            and proc_info["status"]
-                            != psutil.STATUS_ZOMBIE  # type: ignore # psutil.STATUS_ZOMBIE exists
-                            and proc_info.get("ppid") == parent_screen_pid
-                            and proc_info.get("cwd")
-                        ):
-                            proc_cwd_normalized = os.path.normpath(
-                                os.path.abspath(proc_info["cwd"])
-                            )
-                            if proc_cwd_normalized == target_server_dir_norm:
-                                pid = proc_info["pid"]
-                                bedrock_psutil_process = psutil.Process(pid)
-                                found_child = True
-                                break
-                    except (psutil.Error, TypeError):
-                        continue
-                if not found_child:
-                    self.logger.debug(
-                        f"bedrock_server process not found as child of screen '{screen_session_name}' in CWD '{target_server_dir_norm}'."
-                    )
-                    return None
-
-            elif self.os_type == "Windows":
-                # bedrock_executable_path is from BaseMixin
-                normalized_expected_exe = os.path.normcase(
-                    os.path.abspath(self.bedrock_executable_path)
-                )
-                found_proc = False
-                for proc_iter in psutil.process_iter(["pid", "name", "exe"]):
-                    try:
-                        proc_info = proc_iter.info
-                        if (
-                            proc_info["name"] == self.bedrock_executable_name
-                        ):  # e.g., "bedrock_server.exe"
-                            proc_exe_path = proc_info.get("exe")
-                            if proc_exe_path:
-                                normalized_proc_exe = os.path.normcase(
-                                    os.path.abspath(proc_exe_path)
-                                )
-                                if normalized_proc_exe == normalized_expected_exe:
-                                    pid = proc_info["pid"]
-                                    bedrock_psutil_process = psutil.Process(pid)
-                                    found_proc = True
-                                    break
-                    except (psutil.Error, TypeError):
-                        continue
-                if not found_proc:
-                    self.logger.debug(
-                        f"Windows process for '{self.bedrock_executable_path}' not found."
-                    )
-                    return None
-            else:
-                self.logger.error(
-                    f"Process info retrieval not supported on OS: {self.os_type}"
-                )
-                return None  # Or raise NotImplementedError
-
-            # --- Get Process Details using psutil AND Calculate Delta CPU ---
-            if bedrock_psutil_process and pid:
-                try:
-                    with bedrock_psutil_process.oneshot():
-                        # --- CPU Delta Calculation ---
-                        current_proc_cpu_times = (
-                            bedrock_psutil_process.cpu_times()
-                        )  # psutil._common.scpustats
-                        current_sample_time = time.time()
-                        cpu_percent = 0.0
-
-                        if (
-                            self._last_proc_cpu_times_stats is not None
-                            and self._last_proc_sample_time is not None
-                        ):
-                            time_interval = (
-                                current_sample_time - self._last_proc_sample_time
-                            )
-                            if (
-                                time_interval > 0.01
-                            ):  # Avoid division by zero or tiny intervals
-                                prev_proc_cpu_times = self._last_proc_cpu_times_stats
-
-                                # Calculate total CPU time used (user + system) by this process since last measurement
-                                process_cpu_time_delta = (
-                                    current_proc_cpu_times.user
-                                    - prev_proc_cpu_times.user
-                                ) + (
-                                    current_proc_cpu_times.system
-                                    - prev_proc_cpu_times.system
-                                )
-
-                                # Calculate percentage over the elapsed time window
-                                cpu_percent = (
-                                    process_cpu_time_delta / time_interval
-                                ) * 100
-                                self.logger.debug(
-                                    f"Delta CPU Calc for '{self.server_name}' (PID={pid}): "
-                                    f"TimeInterval={time_interval:.3f}s, CPUTimeDelta={process_cpu_time_delta:.4f}, CPU%={cpu_percent:.1f}"
-                                )
-                            else:
-                                self.logger.debug(
-                                    f"Delta CPU Calc for '{self.server_name}' (PID={pid}): Interval too small ({time_interval:.3f}s). Reporting previous or 0%."
-                                )
-                        else:
-                            self.logger.debug(
-                                f"Delta CPU Calc for '{self.server_name}' (PID={pid}): No previous data. Reporting 0.0% CPU for first measurement."
-                            )
-
-                        # Update instance state for the next call for this server
-                        self._last_proc_cpu_times_stats = current_proc_cpu_times
-                        self._last_proc_sample_time = current_sample_time
-                        # --- End CPU Delta Calculation ---
-
-                        memory_info = bedrock_psutil_process.memory_info()
-                        memory_mb = memory_info.rss / (
-                            1024 * 1024
-                        )  # Resident Set Size in MB
-
-                        create_time = bedrock_psutil_process.create_time()
-                        uptime_seconds = current_sample_time - create_time
-                        uptime_str = str(timedelta(seconds=int(uptime_seconds)))
-
-                        process_info_dict = {
-                            "pid": pid,
-                            "cpu_percent": round(cpu_percent, 1),
-                            "memory_mb": round(memory_mb, 1),
-                            "uptime": uptime_str,
-                        }
-                        self.logger.debug(
-                            f"Retrieved process info for '{self.server_name}': {process_info_dict}"
-                        )
-                        return process_info_dict
-
-                except (
-                    psutil.NoSuchProcess,
-                    psutil.AccessDenied,
-                    psutil.ZombieProcess,
-                ) as e_ps_detail:
-                    self.logger.warning(
-                        f"Process PID {pid} for '{self.server_name}' disappeared or access denied during detail retrieval: {e_ps_detail}"
-                    )
-                    # Clean up stale PID specific CPU times if process disappears
-                    self._last_proc_cpu_times_stats = None
-                    self._last_proc_sample_time = None
-                    return None  # Consistent with original behavior
-                except Exception as detail_err:  # Other psutil errors or unexpected
-                    self.logger.error(
-                        f"Error getting details for process PID {pid} ('{self.server_name}'): {detail_err}",
-                        exc_info=True,
-                    )
-                    # Original _get_bedrock_process_info raised ResourceMonitorError here.
-                    raise ResourceMonitorError(
-                        f"Error getting process details for '{self.server_name}' (PID {pid}): {detail_err}"
-                    ) from detail_err
-            else:  # bedrock_psutil_process or pid is None after search
-                self.logger.debug(
-                    f"Bedrock process object not obtained for '{self.server_name}'. Cannot get info."
-                )
-                return None
-
-        except (
-            ResourceMonitorError
-        ):  # Re-raise if explicitly raised by us (e.g. psutil not available)
-            raise
-        except (
-            Exception
-        ) as e_find:  # Catch unexpected errors during the process finding stage
-            self.logger.error(
-                f"Unexpected error finding process for '{self.server_name}': {e_find}",
-                exc_info=True,
-            )
-            # Original _get_bedrock_process_info raised ResourceMonitorError here.
-            raise ResourceMonitorError(
-                f"Unexpected error finding process info for '{self.server_name}': {e_find}"
-            ) from e_find
+        return system_base._get_bedrock_process_info(self.server_name, self.base_dir)
